Refactor the interface
This change has two purpose:
1) add priceFeedExists to handle the case when a price does not exist
in `updatePriceIfNecessary`.
try-catch doesn't work on local function call in solidity and we cannot
catch the revert there.
2) getValidTimePeriod(): There are some cases that the previous
getCurrentPrice implementation does not
cover and it is better to use the latestAvailablePrice.
However, we need to know the validity period and this is why it is added
here. It uses the unsafe function to have better revert message if it
occurs.